requirements.txt
python-dotenv
roboflow
tensorflow==2.20.0
scikit-image==0.25.2
pillow==12.0.0